requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
ldap3==2.9.1
python-dotenv==1.0.0
pydantic==2.5.2